ml/requirements.txt
# Core ML frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Model architectures
timm>=0.9.12  # FastViT, EfficientViT, etc.
ultralytics>=8.0.0  # YOLOv8

# Metric learning
pytorch-metric-learning>=2.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.0

# Vector search
annoy>=1.17.0

# Model optimization
onnx>=1.14.0
onnxruntime>=1.16.0
coremltools>=7.0  # iOS export

# Training utilities
lightning>=2.1.0
wandb>=0.16.0
tensorboard>=2.14.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0
typer>=0.9.0
